+++ oak_functions/client/java/src/com/google/oak/functions/client/AttestationClient.java
@@ -18,7 +18,6 @@ package com.google.oak.functions.client;
 
 import com.google.common.base.VerifyException;
 import com.google.oak.remote_attestation.Attestor;
-import com.google.oak.remote_attestation.Encryptor;
 import com.google.protobuf.ByteString;
 import com.google.protobuf.InvalidProtocolBufferException;
 import io.grpc.ManagedChannel;
@@ -47,7 +46,7 @@ public class AttestationClient {
     private final ManagedChannel channel;
     private final StreamObserver<AttestedInvokeRequest> requestObserver;
     private final BlockingQueue<AttestedInvokeResponse> messageQueue;
-    private final Encryptor encryptor;
+    private final Attestor.AttestedPeer attestedPeer;
 
     public AttestationClient(String uri) throws GeneralSecurityException, InterruptedException {
         // Create gRPC channel.
@@ -82,7 +81,7 @@ public class AttestationClient {
         requestObserver = stub.attestedInvoke(responseObserver);
 
         // Generate client private/public key pair.
-        Attestor attestor = new Attestor();
+        Attestor.UnattestedPeer attestor = new Attestor.UnattestedPeer();
         byte[] publicKey = attestor.getPublicKey();
         ByteString publicKeyBytes = ByteString.copyFrom(publicKey);
 
@@ -97,7 +96,7 @@ public class AttestationClient {
         // Remotely attest peer.
         byte[] peerPublicKey = response.getServerIdentity().getPublicKey().toByteArray();
         byte[] peerAttestationInfo = response.getServerIdentity().getAttestationInfo().toByteArray();
-        encryptor = attestor.attest(peerPublicKey, peerAttestationInfo);
+        attestedPeer = attestor.attest(peerPublicKey, peerAttestationInfo);
     }
 
     private Boolean verifyAttestation(byte[] attestationInfo) {
@@ -115,7 +114,7 @@ public class AttestationClient {
     @SuppressWarnings("ProtoParseWithRegistry")
     public Response send(Request request)
             throws GeneralSecurityException, InterruptedException, InvalidProtocolBufferException {
-        byte[] encryptedMessage = encryptor.encrypt(request.getBody().toByteArray());
+        byte[] encryptedMessage = attestedPeer.encrypt(request.getBody().toByteArray());
         oak.functions.server.Request serverRequest =
                 oak.functions.server.Request.newBuilder()
                         .setEncryptedPayload(ByteString.copyFrom(encryptedMessage))
@@ -127,7 +126,7 @@ public class AttestationClient {
         AttestedInvokeResponse attestedResponse = messageQueue.take();
 
         byte[] responsePayload = attestedResponse.getEncryptedPayload().toByteArray();
-        byte[] decryptedResponse = encryptor.decrypt(responsePayload);
+        byte[] decryptedResponse = attestedPeer.decrypt(responsePayload);
         return Response.parseFrom(decryptedResponse);
     }
 }
