--- remote_attestation/java/src/com/google/oak/remote_attestation/Encryptor.java
@@ -1,37 +1,0 @@
-//
-// Copyright 2021 The Project Oak Authors
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-//
-
-package com.google.oak.remote_attestation;
-
-import com.google.oak.remote_attestation.crypto.AeadEncryptor;
-import java.security.GeneralSecurityException;
-
-/** Wrapper for `com.google.oak.remote_attestation.AeadEncryptor`. */
-public class Encryptor {
-    private final AeadEncryptor encryptor;
-
-    protected Encryptor(AeadEncryptor encryptor) {
-        this.encryptor = encryptor;
-    }
-
-    public byte[] encrypt(byte[] data) throws GeneralSecurityException {
-        return encryptor.encrypt(data);
-    }
-
-    public byte[] decrypt(byte[] data) throws GeneralSecurityException {
-        return encryptor.decrypt(data);
-    }
-}
